\name{thinnedDDP-package}
\alias{thinnedDDP-package}
\alias{thinnedDDP}
\docType{package}
\title{
\packageTitle{thinnedDDP}
}
\description{
\packageDescription{thinnedDDP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{thinnedDDP}
\packageIndices{thinnedDDP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{thinnedDDP}

Maintainer: \packageMaintainer{thinnedDDP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
